pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
django-filter = "^22.1"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
psycopg2-binary = "^2.9.5"
django-allauth = "^0.52.0"
django-rest-auth = "^0.9.5"
serializers = "^0.2.4"
channels = "^4.0.0"
django-debug-toolbar = "^3.8.1"
[tool.poetry.dev-dependencies]
